Fixed error while submit quiz flashcard

+++ php/classes/class-qmn-quiz-manager.php
@@ -2137,8 +2137,8 @@ class QMNQuizManager {
 						$results_array = apply_filters( 'qmn_results_array', $results_array, $question );
 						// If question was graded correctly.
 						if ( ! isset( $results_array['null_review'] ) ) {
-							$points_earned += (int)$results_array['points'];
-							$answer_points += (int)$results_array['points'];
+							$points_earned += (float)$results_array['points'];
+							$answer_points += (float)$results_array['points'];
 
 							// If the user's answer was correct.
 							if ( isset( $results_array['correct'] ) && ( 'correct' == $results_array['correct'] ) ) {
